refactor: extract RecordingSessionStopFailurePresenter.swift from RecordingSessionPresenters.swift
Splits the `RecordingSessionStopFailurePresenter` (~36 lines) out of
RecordingSessionPresenters.swift into its own file.

Byte-identical move of lines 154-189. Both files import Combine +
Foundation.

RecordingSessionPresenters.swift: 190 → 153 lines. Tests: 667.
